Fix escaped like wildcards in `like_utf8` / `nlike_utf8` kernels (#2258)
* Fix escaped like wildcards

Added a new function that replaces the like wildcards '%' and '_' for
the regex counterparts before executing them. It also takes into account
that the wildcards can be escaped, in that case, it does remove the
escape characters and leaves the wildcards so that they are matched
against the raw character.

This is implemented iterating over all the characters of the pattern to
figure out when it needs to be transformed or not.

* Rewrite logic with peek after PR feedback

* Simplifly logic

* Add documentation and refactor string creation in tests

* Add small fix and cargo fmt