feat: Implement Parquet footer-mode encryption
Uses PARC magic.  Haven't tested if we perfectly follow the Parquet spec.

Includes some design choices around key metadata (such as using Sha3
key metadata to select keys) which are not sufficiently abstracted for
a general purpose library.  (That is, there would be some changes to
make if this were to be upstreamed.)